fix(filament,gltfio): wasm runtime marshaling
- Engine.getSteadyClockTimeNano: Filament.js returns a BigInt; coerce it to a
  number in JS before it crosses the interop boundary. wasmJs's adapter throws
  marshaling a BigInt into the externally-typed Double (js silently tolerates it).
- gltfio FilamentAsset: drop redundant .toFloat() now that the read helper
  returns a FloatArray; tidy AssetLoader imports.

+++ kotlin/gltfio/src/webMain/kotlin/io/github/erkko68/filament/gltfio/FilamentAsset.js.kt
@@ -62,12 +62,12 @@ actual class FilamentAsset(
         val minArr = aabb.min!!.toFloatArray(3)
         val maxArr = aabb.max!!.toFloatArray(3)
         return Box(
-            (minArr[0].toFloat() + maxArr[0].toFloat()) / 2f,
-            (minArr[1].toFloat() + maxArr[1].toFloat()) / 2f,
-            (minArr[2].toFloat() + maxArr[2].toFloat()) / 2f,
-            (maxArr[0].toFloat() - minArr[0].toFloat()) / 2f,
-            (maxArr[1].toFloat() - minArr[1].toFloat()) / 2f,
-            (maxArr[2].toFloat() - minArr[2].toFloat()) / 2f
+            (minArr[0] + maxArr[0]) / 2f,
+            (minArr[1] + maxArr[1]) / 2f,
+            (minArr[2] + maxArr[2]) / 2f,
+            (maxArr[0] - minArr[0]) / 2f,
+            (maxArr[1] - minArr[1]) / 2f,
+            (maxArr[2] - minArr[2]) / 2f
         )
     }
